requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
opencv-python-headless==4.9.0.80
ultralytics==8.3.127
requests==2.31.0
numpy==1.26.4
google-generativeai==0.5.2
python-multipart==0.0.9
